admin dashboard uncompleted

+++ src/main/java/nus/iss/gdipsa/team7/service/DashboardService.java
@@ -3,9 +3,9 @@ package nus.iss.gdipsa.team7.service;
 public interface DashboardService {
 	
 	long getTotalGamers();
-    int getTotalDevelopers();
-    int getTotalGames();
-    int getGamesPendingReview();
-    int getUserReports();
+    long getTotalDevelopers();
+    long getTotalGames();
+    long getGamesPendingReview();
+    long getUserReports();
 
 }
